:recycle: Remove old box component and rename BoxV2 and update import
