+++ i18n/locales/en/common.json
@@ -77,7 +77,6 @@
   "Choose the right words.": "Choose the right words.",
   "Choose your avatar": "Choose your avatar",
   "Clear All filters": "Clear All filters",
-  "Coming soon.": "Coming soon.",
   "Community feed": "Community feed",
   "Confirm": "Confirm",
   "Confirm on {{deviceModel}}": "Confirm on {{deviceModel}}",
@@ -226,7 +225,6 @@
   "Learn about blockchain with our comprehensive knowledge base.": "Learn about blockchain with our comprehensive knowledge base.",
   "Learn more": "Learn more",
   "Learn more about Lisk ID initialization": "Learn more about Lisk ID initialization",
-  "Learn more about Lisk sidechains": "Learn more about Lisk sidechains",
   "Ledger Nano S Connected.": "Ledger Nano S Connected.",
   "Ledger Nano S Disconnected.": "Ledger Nano S Disconnected.",
   "Like almost all blockchains Lisk has transparency at it’s core. This means that anyone can view everything that happens on the Lisk network, including the holdings of each account.": "Like almost all blockchains Lisk has transparency at it’s core. This means that anyone can view everything that happens on the Lisk network, including the holdings of each account.",
@@ -384,7 +382,6 @@
   "Show passphrase": "Show passphrase",
   "Show the QR code": "Show the QR code",
   "Show this account's transactions on the dashboard.": "Show this account's transactions on the dashboard.",
-  "Sidechains will revolutionize the way decentralized apps are developed. Here you will be able to find hosts, and monitor your sidechains soon.": "Sidechains will revolutionize the way decentralized apps are developed. Here you will be able to find hosts, and monitor your sidechains soon.",
   "Sign In": "Sign In",
   "Sign a message": "Sign a message",
   "Sign back in": "Sign back in",
