--- src/components/boxV2/box.css
@@ -1,61 +1,0 @@
-@import '../app/variablesV2.css';
-@import '../app/mixins.css';
-
-.wrapper {
-  align-items: inherit;
-  background: var(--color-white);
-  border: 1px solid var(--color-platinum);
-  border-radius: var(--border-radius-box);
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  height: auto;
-  overflow: hidden;
-  width: 100%;
-}
-
-.withHeader {
-  margin-bottom: 12px;
-
-  & > header {
-    align-items: center;
-    background-color: var(--color-white-smoke);
-    border-bottom: 1px solid var(--color-platinum);
-    border-top-left-radius: var(--border-radius-box);
-    border-top-right-radius: var(--border-radius-box);
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    height: 48px;
-    justify-content: space-between;
-    margin: 0;
-    padding: 16px 20px;
-    width: 100%;
-
-    & h1,
-    & h2 {
-      @mixin contentLargest bold;
-
-      color: var(--color-maastricht-blue);
-      margin: 0;
-      padding: 0;
-    }
-
-    & input {
-      min-width: 177px;
-    }
-  }
-
-  & > section {
-    margin: 0 20px;
-    padding: 20px 0;
-
-    &:not(:last-child) {
-      border-bottom: 1px solid var(--color-platinum);
-    }
-  }
-
-  & > p {
-    color: var(--color-blue-gray);
-  }
-}
