--- src/components/boxV2/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import styles from './box.css';
-
-const Box = (props) => {
-  const childs = props.children;
-
-  const hasHeader = Array.isArray(childs) && childs.some(child => child && child.type === 'header');
-
-  return (
-    <div className={`${styles.wrapper} ${hasHeader ? styles.withHeader : ''} ${props.className}`}>
-      { props.children }
-    </div>
-  );
-};
-
-export default Box;
