+++ src/components/help/help.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
-import Box from '../boxV2';
+import Box from '../box';
 import { PrimaryButtonV2 as Button } from '../toolbox/buttons/button';
 import styles from './help.css';
 import links from '../../constants/externalLinks';
