+++ src/components/votes/votesTab.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import { translate } from 'react-i18next';
-import BoxV2 from '../boxV2';
+import Box from '../box';
 import AccountVisual from '../accountVisual';
 import VotesTableHeader from './votesTableHeader';
 import TableRow from '../toolbox/table/tableRow';
@@ -92,7 +92,7 @@ class VotesTab extends React.Component {
       || this.state.isLoading;
 
     return (
-      <BoxV2 className={`${styles.wrapper}`}>
+      <Box className={`${styles.wrapper}`}>
         <header>
           <h1>{t('Voted delegates')}</h1>
           <div className={`${styles.filterHolder}`}>
@@ -178,7 +178,7 @@ class VotesTab extends React.Component {
           )
           }
         </main>
-      </BoxV2>
+      </Box>
     );
   }
 }
