+++ src/components/wallet/balanceChart.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { translate } from 'react-i18next';
 import { Line as LineChart } from 'react-chartjs-2';
-import BoxV2 from '../boxV2';
+import Box from '../box';
 import styles from './balanceChart.css';
 import * as ChartUtils from '../../utils/balanceChart';
 import { tokenMap } from '../../constants/tokens';
@@ -22,7 +22,7 @@ class BalanceGraph extends React.Component {
     });
 
     return (
-      <BoxV2 className={`${styles.wrapper}`}>
+      <Box className={`${styles.wrapper}`}>
         <header>
           <h1>{t('{{token}} Balance', { token: tokenMap[token].label })}</h1>
         </header>
@@ -34,7 +34,7 @@ class BalanceGraph extends React.Component {
             />
           </div>
         </main>
-      </BoxV2>
+      </Box>
     );
   }
 }
